+++ changelog.md
@@ -33,6 +33,7 @@
 - new version number format (e.g. v2.9.2) is standard in communication w/ io-drivers, old format is deprecated
 - menu button for German "Kurzanleitung" (separately available on github.com/smartvisu-newstuff/kurzanleitung) integrated in system menu
 - language is switched to "en" by default if configured language file does not exist any more
+- new twig function 'twig_items()' provides a list of all items from 'masteritem.json'
 
 ### Improvements
 - function Date().duration changed to accept negative values (for plots reaching into the future)